requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.18
numpy>=1.24.3
customtkinter>=5.2.0
pillow>=10.0.0
